// et-web-app/src/app/dashboard/interfaces/dictInterface.ts
export interface Dict{   
    startToggleYear1?: boolean;   // 2018
    continueToggleYear1?: boolean;
    viewDetailsToggleYear1?: boolean;

    startToggleYear2?: boolean;   // 2017
    continueToggleYear2?: boolean;
    viewDetailsToggleYear2?: boolean;

    startToggleYear3?: boolean;   // 2016
    continueToggleYear3?: boolean;
    viewDetailsToggleYear3?: boolean;

    startToggleYear4?: boolean;   // 2015
    continueToggleYear4?: boolean;
    viewDetailsToggleYear4?: boolean;

    startToggleYear5?: boolean;   // 2014
    continueToggleYear5?: boolean;
    viewDetailsToggleYear5?: boolean;

    startToggleYear6?: boolean;   // 2013
    continueToggleYear6?: boolean;
    viewDetailsToggleYear6?: boolean;

    startToggleYear7?: boolean;   // 2012
    continueToggleYear7?: boolean;
    viewDetailsToggleYear7?: boolean;

    startToggleYear8?: boolean;   // 2011
    continueToggleYear8?: boolean;
    viewDetailsToggleYear8?: boolean;
}