// et-web-app/src/app/dashboard/interfaces/statusInterface.ts
export interface Status{
    notStartedYear1?: boolean;    // 2018年不存在
    incompleteYear1?: boolean;
    completedYear1?: boolean;

    notStartedYear2?: boolean;    // 2017当前年-1 (没有当前年2018)
    incompleteYear2?: boolean;
    completedYear2?: boolean;

    notStartedYear3?: boolean;
    incompleteYear3?: boolean;
    completedYear3?: boolean;

    notStartedYear4?: boolean;
    incompleteYear4?: boolean;
    completedYear4: boolean;

    notStartedYear5?: boolean;
    incompleteYear5?: boolean;
    completedYear5?: boolean;

    notStartedYear6?: boolean;
    incompleteYear6?: boolean;
    completedYear6?: boolean;

    notStartedYear7?: boolean;
    incompleteYear7?: boolean;
    completedYear7?: boolean;

    notStartedYear8?: boolean;    // 2011最早一年
    incompleteYear8?: boolean;
    completedYear8?: boolean;
}